fix(client-jersey): Convert AddRequestHeaderFilter into an interface
Jersey does not support to add multiple filters of the same type. The
previous implementation was designed to create instances for different
type of headers. The definition as interface enforces to create new
types. New Anonymous implementations are supported by Jersey. The
interface is not designed as functional interface because the Jersey
APIs would require to cast the lambda expressions to a specific type
which works against the readibility of lambdas.