feat(server-mongo-testing): Allow setting custom download path for embedded mongo
The new environment variable EMBEDDED_MONGO_DOWNLOAD_PATH allows to configer the download location of the Mongod executable in case the download from the web is undesired.
I didn't added tests, because testing is difficult without slowing down the other tests. We would need to clear the cache where the mongod executable is stored, so that it can be downloaded every time - which is exactly the behavior which we try to avoid.