fix(dependencies): Upgrade javax.ws.rs to drop a JavaEE API 8 dependency
In version 2.0.1 the Response did not implement AutoClosable although it
has a close() method and has to be closed after use. This is fixed in
version 2.1.1. Therefore we had JavaEE dependencies providing the same
classes with different implementation and were lucky that they won
against the javax.ws.rs dependency.