fix(legal): Replace JEE EL API from JBoss with the original from javax
The JBoss library contains additional classes by JBoss that do not
contain the "Classpath exception" in their legal header.

Therefore we use the original spec now. There are small differences in
the implementations but they should not affect the usage of the API
unless Weld uses the additional classes somewhere outside the API
module.