fix: fix that the GET request after a SEE OTHER response of a POST request is broken
When using a POST request to create an update task, as suggested by the REST guide, a SEE OTHER status with a location is returned. If the jersey client is used that automatically follows redirects (the default), an error occures. Depeding on the server this is either a 400 BAD REQUEST (Dropwiard) or a 500 INTERNAL SERVER ERROR (Wiremock). This error is caused by the fact that the json entity in the POST request is compressed usaing gzip and the header Content-Encoding is set to gzip. As the following GET request is executed with exactly the same headers it is also send with Content-Encoding, however a GET request has no body. Both Dropwizard and Wiremock (both using Jetty as a http server) are unable to handle this kind of malformed request.
As we couldn't find another option to disable this behavior, e.g. by removing the headers from the GET request (all filters only see this request pair as one request), or fixing the bug in Dropwizard and Wiremock (very deep internals), we decided that the currently the only option is to disable request body compression.
Therefor, by default, request body compression is disabled for the jersey client. The impact shouldn't be to big, as compression for responses is still active.
If one still want to use request body compression, it is now possible to activate it via the client builder.
During these changes we also noticied that it might be necessary in some cases to give an API client, build from a proxy interface, a name. This is required as Dropwizard expect every client instance to have a unique name (e.g. used for metrics and executor service).