fix(server-auth): allow to add jwkskey without algorithm parameter
The "alg" (algorithm) parameter identifies the algorithm intended for
use with the key.  The values used should either be registered in the
IANA "JSON Web Signature and Encryption Algorithms" registry
established by [JWA] or be a value that contains a Collision-Resistant Name.
The "alg" value is a case-sensitive ASCII string.
Use of this member is OPTIONAL.