feat(server-auth): add configuration option to disable the default annotation based authorization by still validating tokens if received
This change modifies the initialization of the AuthBundle as well as
the SdaPlatformBundle.

The change in the AuthBundle.builder() is _breaking_ since it requires
the selection of an Authorization policy. Beside the previous (default)
behavior, a new strategy is added to allow full external authorization.

Old syntax:
    ...
    AuthBundle.builder().withAuthConfigProvider(AuthTestConfig::getAuth).build()
    ...

New syntax to keep the old behavior:
    ...
    AuthBundle.builder().withAuthConfigProvider(AuthTestConfig::getAuth).withAnnotatedAuthorization().build()
    ...

The new option `.withExternalAuthorization()` removes the need for the
`@PermitAll` annotation and enables the token validation for all
endpoints. However, it does not require a token to be present since
this decision is moved to the authorizer (e.g. the Open Policy Agent).

The change in the SdaPlatformBundle.builder() is non-breaking. It adds
the possibility to install the OpaBundle as advanced configuration
option. This also configures the AuthBundle with the
`.withExternalAuthorization()` setting.

Integrate the OpaBundle:
    ...
    SdaPlatformBundle<MyConfiguration> bundle = SdaPlatformBundle.builder()
          .usingCustomConfig(MyConfiguration.class)

          // Configure both the AuthBundle and the OpaBundle
          .withOpaAuthorization(MyConfiguration::getAuth, MyConfiguration::getOpa)

          .withoutCorsSupport(MyConfiguration::getCors)
          ...
          .build();
    ...