feat(server-auth): move the constraint definition from being a single rule to interpreting all top-level rules as constraints
This change requires changes in the policy, while the Java interface remains unchanged.

Previously, a valid OPA policy answer with constraints looked like this:
```
{
  "result": {
    "allow": true,
    "constraints": {
      "agent_ids": ["122", "1234"],
      "customer_ids": ["456", "654"],
      "full_access": ["true"]
    }
  }
}
```

Now this is changed to this:
```
{
  "result": {
    "allow": true,
    "agent_ids": ["122", "1234"],
    "customer_ids": ["456", "654"],
    "full_access": ["true"]
  }
}
```

Note that the constraints are no longer properties of a single rule but instead each top-level rule represents a constraint.

Co-authored-by: Oliver Sand <oliver.sand@sda-se.com>