feat(client-jersey): Enable usage of default methods in client interfaces
With this feature default methods can be used in client interfaces that
are created for the Jersey Proxy API.

This can simplify the API of POST methods by using method parameters for
the properties of a JSON to send. The actual JSON object must not be
created in the client code but can be created in the interface.

Error handling can be implemented in the interface instead of creating
an additional client service class.

Also resilience patterns like retry or circuit breaker may be
implemented without delegating through multiple services.