fix(server-mongo): retry starting the embedded mongo db if exceptions occur during download of the binary
If multiple tests that use the MongoDbRule run in parallel on a fresh machine, the downloads can collide leading to failing tests. This commit adds error handling that retries the initialization and start of the embedded mongo up to 10 times to overcome this issue. This change is a workaround until the used flapdoodle embedded mongodb provides a fix for this.