feat(server-auth-testing): add more flexible request matching to the OPA rule
This change adds support for checking whether the JWT was correctly sent to the OPA. In the old implementation one could use `onRequest(method, path)` to configure an expected request. Now there is a builder approach `onRequest().withMethod(method).withPath(path).withJwt(jwt)` that also allows to check for explicit JWT values. This is also possible for verifying requests: `verify(1, onRequest().withMethod(method).withPath(path).withJwt(jwt))`. There was no functional change in the existing methods.

The existing `onRequest(method, path)` approach was also marked as deprecated since the new approach gives much more flexibility. Futhermore, it should be advised to always check if the JWT is if correctly forwarded to the OPA.