feat(auth): Provide `OpaJwtPrincipal` as injectable bean
In many of our services we add a dedicated class or method that extracts and
casts the principal from the injected `SecurityContext`.

To avoid the repeated block of code that always has to be tested, we can
inject the `OpaJwtPrincipal` directly to read the constraints defined by
the OPA.

Unfortunately there seems to be no option to provide constraints in the
context as they depend on the project.