fix(deps): fix overriding the kotlin dependecy (and downgrade to 1.3.50)
Constraints are evaluated before included BOMs which means that we can't override dependencies included by BOMs with a constraint later. The previous comment said that, but didn't worked and caused mixed versions and a downgrade to a very old version that contains CVEs. This can be fixed by including the kotlin-bom, that also makes sure that all depending modules stay on the same version.
However this doesn't work with 1.3.72 for some reason, so we downgrade to 1.3.50 for now.