fix(server-security): always enable useForwardedHeaders by default
The default value was changed in https://github.com/dropwizard/dropwizard/pull/2748. Our services are expected to be run in a Kubernetes-based environment where we expect that we always have a reverse-proxy or load-balancer. Before, we raised an exception when this option was deactivated. To not burden each service to change this setting we decided to change it here to our expected default.