fix(deps): upgrade hamcrest
hamcrest (2.x) and hamcrest-core (1.x) have mostly similar classes
but are not compatible in all details the later version of hamcrest
comes with Awaitility, so we use that in favour of the older one
served with junit. It is assumed that junit does not use it itself
and therefore has no conflicts due to the changes.

As junit by default provides hamcrest, we still ship it. But it is
not used any more here in favour of AssertJ. All usages of hamcrest
are replaced with AssertJ.

We need to manage the deprecated dependency hamcrest-core to
overwrite transitive dependencies. Version 2.2 of hamcrest-core
contains deprecation notices only and has no conflicts with the new
artifact.