fix(server-morphia): Remove deprecated methods
BREAKING CHANGE:
`MorphiaBundle` can not be initialized without `LocalDateConverter` any more.
Consumers storing LocalDate in MongoDB without a converter or a custom converter that produces a different format need to upgrade their collections. If services use no converter or a custom converter for `LocalDate` in MongoDB entities the collection has to be migrated. No action must be taken if `LocalDate` is not used for properties in MongoDB entities. No action must be taken of PostgreSQL/JPA is used to persist data.