feat: set Dropwizard 1.x.x in maintenance mode
BREAKING CHANGE:

This is the last regular release of sda-dropwizard-commons with support for Dropwizard 1.x.x. Security updates will be applied for a while to give consumers time for an upgrade.

Future releases of sda-dropwizard-commons depend on Dropwizard 2.x.x. We use the major release of Dropwizard to clean up some deprecations and remove support of some unused technologies. Please find all information about removed features and APIs in the deprecation notes below.

# Deprecations

## Java 8

Although not decided finally, it is planned to drop support for Java 8 with the upgrade to Dropwizard 2 or within the lifecycle of Dropwizard 2 in sda-dropwizard-commons. We highly recommend to upgrade dependent services to Java 11 before upgrading to sda-dropwizard-commons with Dropwizard 2. This recommendation may change to a requirement.

## Avro and Confluent

Avro is a binary format from Confluent to specify the payload of Kafka messages in a backward compatible way. We came to the conclusion that it causes more trouble than it helps us. All SDA services have changed to JSON for the message payload. Avro causes additional effort and care in the dependency management in combination with Kafka.

Avro will no longer be supported by sda-dropwizard-commons and the modules `sda-commons-server-kafka-confluent` and `sda-commons-server-kafka-confluent-testing` will be removed when upgrading to Dropwizard 2.

## Weld/CDI

Weld is used for dependency injection. It makes the microservices feel more like JEE applications and aligns with other JEE standards like the used annotations for HTTP endpoints. On the other hand, we had a lot of trouble with Weld during development. Additional classpath scanning and proxy creation slows down application startup.

The standard pattern of SDA SE is to avoid Weld and use HK2 to inject context information into endpoint implementations. Instances are created with `new`.

As a reminder we like to announce that SDA SE teams will not put much effort into enhancements of the Weld module. Contributions are welcome. Weld is NOT deprecated in sda-dropwizard-commons.

## sda-commons-client-jersey

Client Jersey has some deprecated methods in the factory. Alternatives are available and documented.

The deprecated methods in Client Jersey will be removed.

## sda-commons-server-morphia

The `MorphiaBundle` has the option to disable the provided `LocalDateConverter` because it was not available from the beginning.

We will remove that option to align how `LocalDate` is stored in MongoDB across services. Consumers storing LocalDate in MongoDB without a converter or a custom converter  that produces a different format, need to upgrade their collections. If services use no converter or a custom converter for `LocalDate` in MongoDB entities the collection has to be migrated. No action must be taken of `LocalDate` is not used for properties in MongoDB entities. No action must be taken of PostgreSQL/JPA is used to persist data.

## sda-commons-server-auth-testing

A method in the `OpaRule` for the `AllowBuilder` that does not match the readable fluent builder pattern is deprecated.

The deprecated method will be removed. The alternative is easy to adopt.

## sda-commons-shared-asyncapi

Annotations have been added to avoid class not found exceptions because Dropwizard 1 uses validation api 1. They will be useless as they come with validation api 2 which is default in Dropwizard 2.

Temporarily added annotations will be removed. No action is needed by consumers. Annotations from the package `javax.validation` can now be used as intended.

## sda-commons-server-kafka

There are some configurations left from the time before `MessageListenerStrategy` has been introduced. `MessageListenerStrategy` is the desired way to configure Kafka Consumers.

We will remove the legacy configuration. Consumers need to change to `MessageListenerStrategy`.