test(server-morphia): fix test to verify that no personal data is traced
IntelliJ claimed that `noneMatch(Tags.DB_STATEMENT::equals)` is comparing incompatible types, thanks for that.
Actually it should have been `noneMatch(Tags.DB_STATEMENT.getKey()::equals)`.
But that is the wrong assertion because we do not remove the statement from the trace but replace it with a sanitized version.

This change asserts that we want the statement without the personal information in the trace.