feat(client-jersey): add dedicated config class for Api clients
In many service we had to extend the HttpClientConfiguration to add a base url of the API which is required for API clients based on interfaces.
This commit provides a built in approach to simplify configuration and creation of API clients.
The downside is that there are some classes partially duplicated due to the builder pattern.
But that gives the opportunity to get rid of the deprecated timeout settings of the previous fluent configuration.