feat(client-jersey): provide annotations to suppress exceptions when using API clients
We realized, that client errors are handled in only two ways very often:

- some errors are mitigated in the same way, weather it is a 404 or a 403
- all other errors are rethrown to produce a 500 for the caller

Therefore we wrote a lot of code to catch, verify and handle exceptions.
This code can be avoided with the new suppression annotations.
Consumers may still handle exceptions for technical reasons, e.g. do a retry on connection errors, but use the suppression feature for business implementation if for example 404 is an expected result.