feat(starter): use optional consumer token by default
With this change, the consumer token for incoming HTTP request is optional by default. Consumer-Token is only used for logging and metrics. It is more or less used for a tracing in a simple way. Technically the consumer token support is removed from the `SdaPlatformBundle` and `SdaPlatformConfiguration`. The `sda-commons-server-consumer` is left as is, because it is used in the `sda-commons-server-starter`.