feat(server-security): add option to configure CSP for serving frontends from a service
proposed release notes:

If a service is serving HTML pages besides the API, this release requires enabling `withFrontendSupport` when configuring the `SdaPlatformBundle` or the `SecurityBundle`. Otherwise, the frontend will not be able to use scripts, styles, images or web fonts.
If a service is only serving APIs, no changes are needed and the service is more secure due to more restrictive CSP header.