fix(server-mongo-testing): define type mapper for compatibility with Morphia
This way we support generic type mapping (so far for properties not for entities) as it was done with Morphia.

**TODO**

As we customize Spring Data a lot here, we should think about a `withMorphiaCompatibility()` switch in the builder to configure Spring Data MongoDB to be compatible with existing databases.
If not set, we should stick to Spring Data MongoDB defaults as that is what users would expect and we gain the opportunity to refactor the whole service to Spring Boot if desired.