fix(server-jackson): write invalid Json when errors happen while serialization
BREAKING CHANGE:

The ObjectMapper will not produce valid Json anymore when an Exception is thrown while serialization.

Such invalid Jsons are the only way to identify a serialisation error on consumer side when e.g. a Json is written directly to the response output stream. Streaming data directly from a source (e.g. a MongoCursor) into the response as chunked encoding is valuable to minimize memory resources but may lead to not identifiable errors (e.g. while mapping from entity model to API model) when the ObjectMapper creates a valid Json after the error when the generator is closed.