test(jackson): update test that shows incompatibilities
BREAKING CHANGE:

This release updates Jackson from 2.13.5 to 2.15.2 to support Snakeyaml 2.0.
This update resolves CVE-2022-1471 in Snakeyaml.
The behavior of Jackson changed when deserializing timestamps in ISO format as Java's `ZonedDateTime`.

`ZonedDateTime` is created on deserialization with a `ZoneId` `"Z"` instead of `"UTC"` now.
The result references the same time with the same time offset, but is not equal in terms of `ZonedDateTime.equals(other)`
This change can affect API integration tests where assertions are done with the deserialized instances.
It does not affect preferred API integration tests that compare at JSON level.